Refactor: consolidate leader election
Deduplicate leader election code for the manager and the operator.

Preparation for HIVE-1736
